# lib/api-zod/src/generated/types/careCard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CareCardUrgencyLevel } from './careCardUrgencyLevel';

export interface CareCard {
  id: string;
  photoEntryId: string;
  generatedText: string;
  routineSteps: string[];
  ingredients: string[];
  urgencyLevel: CareCardUrgencyLevel;
  createdAt: Date;
}