# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CareCard } from './careCard';
import type { PhotoEntry } from './photoEntry';

export interface ScanResult {
  photoEntry: PhotoEntry;
  careCard: CareCard;
}